backend/requirements.txt
# Core Flask Dependencies
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
python-dotenv==1.1.1

# Database Dependencies
SQLAlchemy==2.0.41
alembic==1.16.4
psycopg2-binary==2.9.10

# Security Dependencies
bcrypt==4.3.0
PyJWT==2.10.1
cryptography==45.0.6
passlib==1.7.4

# Caching and Performance
redis==6.4.0
async_timeout==5.0.1
Flask-Limiter==3.12

# Email and Notifications
email-validator==2.2.0

# Data Processing and Export
openpyxl==3.1.5
pandas==2.3.1
numpy==2.3.2

# System Monitoring
psutil==7.0.0

# Logging and Monitoring
structlog==25.4.0

# Production Server
gunicorn==21.2.0

# Error Tracking (Optional)
sentry-sdk[flask]==1.40.6

# Background Tasks
celery==5.3.4

# Core Python Dependencies
click==8.2.1
blinker==1.9.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
typing_extensions==4.14.0

# Additional Dependencies
Mako==1.3.10
greenlet==3.2.10
dnspython==2.7.0
idna==3.10
et-xmlfile==2.0.0
pytz==2025.2
python-dateutil==2.9.0.post0
tzdata==2025.2
six==1.17.0
packaging==25.0
deprecated==1.2.18
wrapt==1.17.3
cffi==1.17.1
pycparser==2.22
rich==13.9.4
markdown-it-py==4.0.0
mdurl==0.1.2
pygments==2.19.2
ordered-set==4.1.0
limits==5.5.0

# Development and Testing (Optional)
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0

